Rename viewset tests
