+++ api/tests/__init__.py
@@ -1,4 +1,4 @@
-from .user_registration_test import UserRegistrationTestCase  # noqa: F401
-from .campus_test import CampusTestCase  # noqa: F401
-from .course_test import CourseTestCase  # noqa: F401
-from .subject_test import SubjectTestCase  # noqa: F401
+from .user_registration_view_test import UserRegistrationTestCase  # noqa: F401
+from .campus_viewset_test import CampusTestCase  # noqa: F401
+from .course_viewset_test import CourseTestCase  # noqa: F401
+from .subject_viewset_test import SubjectTestCase  # noqa: F401
