+++ api/tests/user_registration_view_test.py
